requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
motor==3.1.1
multidict==6.0.4
python-dotenv==0.21.0
soupsieve==2.3.2.post1
yarl==1.8.2
